requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.9.0
eventlet==0.35.1
Werkzeug==2.3.8
PyJWT==2.8.0
